+++ packages/vue/storybook/.storybook/main.js
@@ -1,7 +1,5 @@
 module.exports = {
   "stories": [
-    // "../stories/**/*.stories.mdx",
-    // "../stories/**/*.stories.@(js|jsx|ts|tsx)",
     "../../**/*.stories.js"
   ],
   "addons": [
@@ -9,15 +7,3 @@ module.exports = {
     "@storybook/addon-essentials"
   ]
 }
-
-// const { configure } = require('@storybook/vue');
-// // automatically import all files ending in *.stories.js
-// const req = require.context(
-//   '../',
-//   true,
-//   /\.stories\.js$/
-// );
-// function loadStories() {
-//   req.keys().forEach(filename => req(filename));
-// }
-// configure(loadStories, module);