+++ packages/css/breadcrumb/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.2.6](https://github.com/Novicell/frontend-packages/compare/@novicell/styles-breadcrumb@0.2.5...@novicell/styles-breadcrumb@0.2.6) (2021-04-12)
+
+**Note:** Version bump only for package @novicell/styles-breadcrumb
+
+
+
+
+
+
 ## [0.2.5](https://github.com/Novicell/frontend-packages/compare/@novicell/styles-breadcrumb@0.2.3...@novicell/styles-breadcrumb@0.2.5) (2021-02-22)